+++ crates/ui/widgets/src/dialog/settings_dialog/tab/general_settings.rs
@@ -1,3 +1,4 @@
+use arama_i18n::Locale;
 use arama_env::target_media_type::TargetMediaType;
 
 pub mod message;
@@ -9,6 +10,7 @@ pub struct GeneralSettings {
     target_media_type: TargetMediaType,
     sub_dir_depth_limit: u8,
     similarity_threshold: f32,
+    locale: Locale,
 }
 
 impl GeneralSettings {
@@ -16,11 +18,13 @@ impl GeneralSettings {
         target_media_type: &TargetMediaType,
         sub_dir_depth_limit: u8,
         similarity_threshold: f32,
+        locale: Locale,
     ) -> Self {
         Self {
             target_media_type: target_media_type.to_owned(),
             sub_dir_depth_limit,
             similarity_threshold,
+            locale,
         }
     }
 }