+++ crates/ui/widgets/src/dialog/settings_dialog/tab/general_settings/message.rs
@@ -1,3 +1,4 @@
+use arama_i18n::Locale;
 use arama_env::target_media_type::TargetMediaType;
 
 #[derive(Debug, Clone)]
@@ -5,4 +6,5 @@ pub enum Message {
     TargetMediaTypeChanged(TargetMediaType),
     SubDirDepthLimitChanged(u8),
     SimilarityThresholdChanged(f32),
+    LocaleChanged(Locale),
 }